# swagger.yaml
swagger: "2.0"
info:
  title: Swagger Chili`s Burger Time
  description: |-
    Building a Game and Website for Chili`s Restaraunts.
    Some useful links:
    - [Repository](https://github.com/Experience-Monks/prj-240137971-chilis-burger-time)
  contact:
    email: developers@mediamonks.com
  version: 1.0.0
host: 3in7ishrah.execute-api.us-east-1.amazonaws.com
basePath: /prod
...
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /leaderboard:
    get:
      tags:
        - Leaderboard
      summary: Returns leaderboard.
      operationId: getLeaderboard
      responses:
        "200":
          description: Top items in the leaderboard.
          schema:
            $ref: "#/definitions/Leaderboard"
        "400":
          description: Bad request - invalid request parameters
        "401":
          description: Unauthorized - authentication failure
        "403":
          description: Forbidden - authorization failure
        "404":
          description: Not found - the session ID does not exist
        "500":
          description: Internal server error
...
definitions:
...
  Score:
    type: object
    properties:
      nickname:
        type: string
        example: jam3
        description: User name to display on the dashboard
      score:
        type: string
        example: 210448
        description: Highest score obtained by the user.
  UserScore:
    type: object
    properties:
      nickname:
        type: string
        example: jam3
        description: User name to display on the dashboard
      score:
        type: string
        example: 210448
        description: Highest score obtained by the user.
      rank:
        type: number
        example: 1
        description: The position on the board
  Scores:
    type: array
    items:
      $ref: "#/definitions/Score"
  Leaderboard:
    type: object
    properties:
      leaderboard:
        $ref: "#/definitions/Scores"
      user:
        $ref: "#/definitions/UserScore"
...
securityDefinitions:
  bearerAuth:
    name: "userToken"
    type: apiKey
    in: header
security:
  - bearerAuth: []